+++ packages/teg-auth/src/machine/messages/delete_task_history.rs
@@ -3,7 +3,7 @@ use teg_protobufs::{
     combinator_message,
 };
 
-pub fn delete_task_history(task_id: u64) -> CombinatorMessage {
+pub fn delete_task_history(task_id: u32) -> CombinatorMessage {
     CombinatorMessage {
         payload: Some(
             combinator_message::Payload::DeleteTaskHistory(
