+++ packages/teg-auth/src/machine/models/speed_controller_r1.rs
@@ -18,7 +18,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(new, Debug, Serialize, Deserialize, Clone)]
 pub struct SpeedController {
-    pub id: u64,
+    pub id: u32,
     // Foreign Keys
     pub component_id: ID,
     // Timestamps
