+++ packages/teg-auth/src/print_queue/tasks/models/task/from_gcodes.rs
@@ -28,7 +28,7 @@ use super::{
 
 impl Task {
     pub async fn from_hook<'ctx>(
-        machine_id: u64,
+        machine_id: u32,
         hook: &String,
         ctx: &'ctx Arc<crate::Context>,
     ) -> Result<Task> {
@@ -40,7 +40,7 @@ impl Task {
     }
 
     pub async fn from_gcodes<'ctx>(
-        machine_id: u64,
+        machine_id: u32,
         gcodes: Vec<String>,
         ctx: &'ctx Arc<crate::Context>,
     ) -> Result<Task> {
